Moved functions to load/interpolate in utils and now lisa calls those
