+++ gw_response/utils.py
@@ -1,8 +1,14 @@
 # Global imports
 import chex
+import h5py
+import interpax
 import jax
 import jax.numpy as jnp
 import jax_healpy as hp
+import numpy as np
+
+# Local imports
+from .constants import PhysicalConstants
 
 # Update jax configuration to enable 64-bit precision for numerical computations
 jax.config.update("jax_enable_x64", True)
@@ -144,3 +150,122 @@ def shift_to_center(first, second, third):
     third_mass = third - center
 
     return first_mass, second_mass, third_mass
+
+
+@jax.jit
+def arms_matrix_from_satellite_positions(m1, m2, m3):
+    """
+    Builds a constellation arm matrix (the vector difference between each
+    ordered pair of satellites) from three satellites' Cartesian positions.
+
+    This differencing is identical regardless of which orbit model produced
+    the positions, so it is shared by the analytical, Keplerian, and
+    numerical orbit models in `gw_response.lisa`.
+
+    Args:
+        m1, m2, m3 (jnp.ndarray): The Cartesian positions of satellites 1, 2,
+        and 3.
+
+    Returns:
+        jnp.ndarray: A numpy array representing the arm matrix of the
+        constellation. Each row of the array corresponds to the vector
+        difference between pairs of satellites, ordered [12, 23, 31, 21, 32,
+        13].
+    """
+    return jnp.array(
+        [
+            m2 - m1,
+            m3 - m2,
+            m1 - m3,
+            m1 - m2,
+            m2 - m3,
+            m3 - m1,
+        ]
+    ).T
+
+
+def _load_numerical_orbits_text(orbit_file):
+    """
+    Loads numerical satellite orbit data from a plain-text file.
+
+    The file is expected to be readable by `numpy.loadtxt` and to contain 10
+    columns: time_in_years, x1, y1, z1, x2, y2, z2, x3, y3, z3, where
+    (xi, yi, zi) are the coordinates (in meters) of satellite i at the given
+    time. One row per time sample, with rows sorted by increasing time.
+    """
+    data = np.atleast_2d(np.loadtxt(orbit_file))
+    if data.shape[1] != 10:
+        raise ValueError(
+            "Numerical orbit files must have 10 columns: time_in_years, x1, "
+            "y1, z1, x2, y2, z2, x3, y3, z3. Got "
+            f"{data.shape[1]} columns instead."
+        )
+    time_grid = jnp.array(data[:, 0])
+    positions_grid = jnp.array(data[:, 1:]).reshape(data.shape[0], 3, 3)
+    return time_grid, positions_grid
+
+
+def _load_numerical_orbits_lisaorbits(orbit_file):
+    """
+    Loads numerical satellite orbit data from an HDF5 orbit file produced by
+    the `lisaorbits` package (https://pypi.org/project/lisaorbits/).
+
+    Only the spacecraft positions (dataset `tcb/x`, shape (size, 3, 3) for
+    (time, satellite, xyz), in meters) and the TCB time grid (attributes
+    `t0` and `dt`, both in seconds, and `size`) are used; velocities,
+    accelerations, light travel times, and pseudoranges are ignored. The
+    time grid is converted from seconds to years to match this module's
+    convention.
+    """
+    with h5py.File(orbit_file, "r") as hdf5:
+        version = str(hdf5.attrs["version"])
+        if int(version.split(".", 1)[0]) < 2:
+            raise ValueError(
+                f"Unsupported lisaorbits file version {version!r}; "
+                "gw_response requires lisaorbits format version >= 2.0."
+            )
+        t0 = float(hdf5.attrs["t0"])
+        dt = float(hdf5.attrs["dt"])
+        size = int(hdf5.attrs["size"])
+        positions_grid = jnp.array(hdf5["tcb/x"][:])
+    time_grid = (t0 + np.arange(size) * dt) / PhysicalConstants().yr
+    return jnp.array(time_grid), positions_grid
+
+
+def load_numerical_orbits(orbit_file, interpolation_method="linear"):
+    """
+    Loads numerical satellite orbit data from an external file and builds an
+    interpolator for the satellite positions.
+
+    Two file formats are supported, auto-detected from the file content:
+      - Plain-text files readable by `numpy.loadtxt`, with 10 columns:
+        time_in_years, x1, y1, z1, x2, y2, z2, x3, y3, z3. See
+        `_load_numerical_orbits_text`.
+      - HDF5 orbit files produced by the `lisaorbits` package
+        (https://pypi.org/project/lisaorbits/), format version >= 2.0. See
+        `_load_numerical_orbits_lisaorbits`.
+
+    The interpolation coefficients are computed once here (similar in spirit
+    to `scipy.interpolate.interp1d`), so evaluating the returned interpolator
+    at query times - even repeatedly inside a jit/vmap - only needs to
+    evaluate the precomputed spline rather than re-deriving it.
+
+    Args:
+        orbit_file (str): Path to the numerical orbit data file.
+        interpolation_method (str, optional): The interpolation method passed
+        to `interpax.Interpolator1D`, e.g. 'linear', 'nearest', 'cubic',
+        'cubic2', 'cardinal', 'catmull-rom', 'monotonic', 'monotonic-0', or
+        'akima'. Default is 'linear'.
+
+    Returns:
+        interpax.Interpolator1D: An interpolator mapping time (in years) to
+        satellite positions. Calling it with query time(s) returns an array
+        of shape (..., 3, 3), indexed as [satellite, coordinate].
+    """
+    if h5py.is_hdf5(orbit_file):
+        time_grid, positions_grid = _load_numerical_orbits_lisaorbits(orbit_file)
+    else:
+        time_grid, positions_grid = _load_numerical_orbits_text(orbit_file)
+    return interpax.Interpolator1D(
+        time_grid, positions_grid, method=interpolation_method
+    )
